refactor(chat): extract chat_manager/{update,errors}.rs (#292)
Moves ProcessedAttachments/ChatUpdate and CommandMeta/SendError/
TrustWorkspaceError/ChatFieldsPartial/ForkError into their own
submodules, re-exported from the module root unchanged.

Task 2 of the chat_manager.rs file-limit split plan.
